Extract latest value logic into useCdaLatestValue
